refactor: inline LazyQueryFS into its test file
Matches the pattern established by ClockFS — backend modules that only
exist to demonstrate the FS.Backend behaviour live alongside their tests,
not in test/support/ under a lib-style namespace.
